Pull common DataDev init code out into common/driver/data_dev_common.c
This code can be shared between data_dev and data_gpu
