+++ data_dev/driver/Makefile
@@ -45,7 +45,7 @@ ccflags-y += -I$(HOME)/src
 ccflags-y += -DDMA_IN_KERNEL=1 -DGITV=\"$(GITV)\"
 
 # Object files that make up the module
-$(NAME)-objs := src/dma_buffer.o src/dma_common.o
+$(NAME)-objs := src/dma_buffer.o src/dma_common.o src/data_dev_common.o
 $(NAME)-objs += src/axi_version.o src/axis_gen2.o src/data_dev_top.o
 
 # Target module to be built
