+++ data_dev/driver/src/data_dev_common.c
@@ -1,0 +1,1 @@
+../../../common/driver/data_dev_common.c
